Refactor binderpos storefront into focused files
