+++ api/gateway/binderpos/storefront_card_helpers.go
@@ -1,0 +1,55 @@
+package binderpos
+
+import (
+	"fmt"
+	"net/url"
+	"strings"
+)
+
+func formatCardName(scrapVariant int, productTitle, variantTitle string) string {
+	productTitle = strings.TrimSpace(productTitle)
+	variantTitle = strings.TrimSpace(variantTitle)
+
+	switch scrapVariant {
+	case 2:
+		if variantTitle == "" {
+			return productTitle
+		}
+		return strings.TrimSpace(productTitle + " - " + variantTitle)
+	case 3:
+		return stripTrailingSet(productTitle)
+	default:
+		return productTitle
+	}
+}
+
+func stripTrailingSet(productTitle string) string {
+	title := strings.TrimSpace(productTitle)
+	open := strings.LastIndex(title, "[")
+	close := strings.LastIndex(title, "]")
+	if open >= 0 && close > open && close == len(title)-1 {
+		return strings.TrimSpace(title[:open])
+	}
+	return title
+}
+
+func extractSetName(productTitle string) string {
+	title := strings.TrimSpace(productTitle)
+	open := strings.LastIndex(title, "[")
+	close := strings.LastIndex(title, "]")
+	if open >= 0 && close > open && close == len(title)-1 {
+		return strings.TrimSpace(title[open+1 : close])
+	}
+	return ""
+}
+
+func buildCardImageURL(rawImageURL, cardTitle string) string {
+	img := strings.TrimSpace(rawImageURL)
+	if strings.HasPrefix(img, "//") {
+		return "https:" + img
+	}
+	if strings.HasPrefix(img, "http://") || strings.HasPrefix(img, "https://") {
+		return img
+	}
+	return fmt.Sprintf("https://placehold.co/304x424?text=%s", url.QueryEscape(strings.TrimSpace(cardTitle)))
+}
